+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/CertificationResult.java
@@ -17,6 +17,7 @@ import org.apache.commons.lang3.StringUtils;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
+import gov.healthit.chpl.conformanceMethod.domain.ConformanceMethod;
 import gov.healthit.chpl.dto.CertificationResultDetailsDTO;
 import gov.healthit.chpl.optionalStandard.domain.CertificationResultOptionalStandard;
 import gov.healthit.chpl.optionalStandard.domain.OptionalStandard;
@@ -167,6 +168,9 @@ public class CertificationResult implements Serializable {
     private String privacySecurityFramework;
 
     @XmlTransient
+    private List<ConformanceMethod> allowedConformanceMethods;
+
+    @XmlTransient
     private List<TestFunctionality> allowedTestFunctionalities;
 
     @XmlTransient
@@ -189,6 +193,13 @@ public class CertificationResult implements Serializable {
     @XmlElement(name = "testFunctionality")
     @Builder.Default
     private List<CertificationResultTestFunctionality> testFunctionality = new ArrayList<CertificationResultTestFunctionality>();
+
+    /**
+     * The conformance method used for the certification criteria
+     */
+    @XmlElementWrapper(name = "conformanceMethods", nillable = true, required = false)
+    @XmlElement(name = "conformanceMethod")
+    private List<CertificationResultConformanceMethod> conformanceMethods= new ArrayList<CertificationResultConformanceMethod>();
 
     /**
      * The test procedures used for the certification criteria
@@ -270,6 +281,7 @@ public class CertificationResult implements Serializable {
         this.testStandards = new ArrayList<CertificationResultTestStandard>();
         this.additionalSoftware = new ArrayList<CertificationResultAdditionalSoftware>();
         this.testDataUsed = new ArrayList<CertificationResultTestData>();
+        this.conformanceMethods = new ArrayList<CertificationResultConformanceMethod>();
         this.testProcedures = new ArrayList<CertificationResultTestProcedure>();
         this.testFunctionality = new ArrayList<CertificationResultTestFunctionality>();
         this.svaps = new ArrayList<CertificationResultSvap>();
@@ -382,6 +394,7 @@ public class CertificationResult implements Serializable {
 
         this.setOptionalStandards(getOptionalStandards(certResult, certRules));
         this.setTestFunctionality(getTestFunctionalities(certResult, certRules));
+        this.setConformanceMethods(getConformanceMethods(certResult, certRules));
         this.setTestProcedures(getTestProcedures(certResult, certRules));
         this.setTestDataUsed(getTestData(certResult, certRules));
         this.setTestToolsUsed(getTestTools(certResult, certRules));
@@ -417,6 +430,16 @@ public class CertificationResult implements Serializable {
         }
     }
 
+    private List<CertificationResultConformanceMethod> getConformanceMethods(CertificationResultDetailsDTO certResult, CertificationResultRules certRules) {
+        if (certRules.hasCertOption(certResult.getNumber(), CertificationResultRules.CONFORMANCE_METHOD)) {
+            return certResult.getConformanceMethods().stream()
+                    .map(item -> new CertificationResultConformanceMethod(item))
+                    .collect(Collectors.toList());
+        } else {
+            return null;
+        }
+    }
+
     private List<CertificationResultTestProcedure> getTestProcedures(CertificationResultDetailsDTO certResult, CertificationResultRules certRules) {
         if (certRules.hasCertOption(certResult.getNumber(), CertificationResultRules.TEST_PROCEDURE)) {
             return certResult.getTestProcedures().stream()
@@ -426,6 +449,7 @@ public class CertificationResult implements Serializable {
             return null;
         }
     }
+
 
     private List<CertificationResultTestData> getTestData(CertificationResultDetailsDTO certResult, CertificationResultRules certRules) {
         if (certRules.hasCertOption(certResult.getNumber(), CertificationResultRules.TEST_DATA)) {
@@ -473,6 +497,14 @@ public class CertificationResult implements Serializable {
 
     public void setId(Long id) {
         this.id = id;
+    }
+
+    public List<CertificationResultConformanceMethod> getConformanceMethods() {
+        return conformanceMethods;
+    }
+
+    public void setConformanceMethods(List<CertificationResultConformanceMethod> conformanceMethods) {
+        this.conformanceMethods = conformanceMethods;
     }
 
     public List<CertificationResultTestProcedure> getTestProcedures() {
@@ -641,6 +673,14 @@ public class CertificationResult implements Serializable {
 
     public void setPrivacySecurityFramework(String privacySecurityFramework) {
         this.privacySecurityFramework = privacySecurityFramework;
+    }
+
+    public List<ConformanceMethod> getAllowedConformanceMethods() {
+        return allowedConformanceMethods;
+    }
+
+    public void setAllowedConformanceMethods(List<ConformanceMethod> allowedConformanceMethods) {
+        this.allowedConformanceMethods = allowedConformanceMethods;
     }
 
     public List<TestFunctionality> getAllowedTestFunctionalities() {