+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/validation/listing/Edition2014ListingValidator.java
@@ -12,6 +12,7 @@ import gov.healthit.chpl.validation.listing.reviewer.CertificationStatusReviewer
 import gov.healthit.chpl.validation.listing.reviewer.ChplNumberComparisonReviewer;
 import gov.healthit.chpl.validation.listing.reviewer.ChplNumberReviewer;
 import gov.healthit.chpl.validation.listing.reviewer.ComparisonReviewer;
+import gov.healthit.chpl.validation.listing.reviewer.ConformanceMethodReviewer;
 import gov.healthit.chpl.validation.listing.reviewer.DeveloperBanComparisonReviewer;
 import gov.healthit.chpl.validation.listing.reviewer.DeveloperStatusReviewer;
 import gov.healthit.chpl.validation.listing.reviewer.DuplicateDataReviewer;
@@ -137,6 +138,10 @@ public abstract class Edition2014ListingValidator extends Validator {
     @Qualifier("optionalStandardReviewer")
     private OptionalStandardReviewer optionalStandardReviewer;
 
+    @Autowired
+    @Qualifier("conformanceMethodReviewer")
+    private ConformanceMethodReviewer conformanceMethodReviewer;
+
     private List<Reviewer> reviewers;
     private List<ComparisonReviewer> comparisonReviewers;
 
@@ -156,6 +161,7 @@ public abstract class Edition2014ListingValidator extends Validator {
             reviewers.add(certDateReviewer);
             reviewers.add(unattestedCriteriaWithDataReviewer);
             reviewers.add(optionalStandardReviewer);
+            reviewers.add(conformanceMethodReviewer);
             reviewers.add(tsrReviewer);
             reviewers.add(ttReviewer);
             reviewers.add(tt2014Reviewer);